% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clean_title}
\alias{clean_title}
\title{Clean Text for Title View}
\usage{
clean_title(txt)
}
\arguments{
\item{txt}{text included}
}
\description{
Clean Text for Title View
}
\examples{
clean_title("muhammad rasyid ridha")
## [1] "Muhammad Rasyid Ridha"
}
