% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{qs_drop}
\alias{qs_drop}
\title{Drop table on SQLite database}
\usage{
qs_drop(tbl_name, force = FALSE,
  db_path = "B:/Cloud/OneDrive/Magnum Opus/project/data/db/db_qs.sqlite")
}
\arguments{
\item{tbl_name}{name of the table}

\item{db_path}{database location}
}
\description{
Drop table on SQLite database
}
