% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{qs_exec}
\alias{qs_exec}
\title{Execute query on SQLite database}
\usage{
qs_exec(query,
  db_path = "B:/Cloud/OneDrive/Magnum Opus/project/data/db/db_qs.sqlite")
}
\arguments{
\item{query}{query statement}

\item{db_path}{database location}
}
\description{
Execute query on SQLite database
}
