% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{qs_glance}
\alias{qs_glance}
\title{Get first 10 rows of table on SQLite database}
\usage{
qs_glance(tbl_name, n = 10,
  db_path = "B:/Cloud/OneDrive/Magnum Opus/project/data/db/db_qs.sqlite")
}
\arguments{
\item{tbl_name}{name of the table}

\item{n}{number of limits}

\item{db_path}{database location}
}
\description{
Get first 10 rows of table on SQLite database
}
