% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{qs_init}
\alias{qs_init}
\title{Initialize SQLite database}
\usage{
qs_init(path = "B:/Cloud/OneDrive/Magnum Opus/project/data/db",
  db_name = "db_qs.sqlite")
}
\arguments{
\item{path}{folder path of the database}

\item{db_name}{name of the database}
}
\description{
Initialize SQLite database
}
