% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{qs_query}
\alias{qs_query}
\title{Read table from SQLite database}
\usage{
qs_query(query,
  db_path = "B:/Cloud/OneDrive/Magnum Opus/project/data/db/db_qs.sqlite")
}
\arguments{
\item{query}{query statement}

\item{db_path}{database location}
}
\description{
Read table from SQLite database
}
