% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{qs_table}
\alias{qs_table}
\title{Show tables on SQLite database}
\usage{
qs_table(pattern = NULL,
  db_path = "B:/Cloud/OneDrive/Magnum Opus/project/data/db/db_qs.sqlite")
}
\arguments{
\item{pattern}{regex pattern}

\item{db_path}{database location}
}
\description{
Show tables on SQLite database
}
