% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{qs_write}
\alias{qs_write}
\title{Write data into table on SQLite database}
\usage{
qs_write(df, tbl_name, append = TRUE, overwrite = FALSE,
  db_path = "B:/Cloud/OneDrive/Magnum Opus/project/data/db/db_qs.sqlite")
}
\arguments{
\item{df}{data.frame to insert}

\item{tbl_name}{name of the table}

\item{append}{TRUE}

\item{overwrite}{FALSE}

\item{db_path}{database location}
}
\description{
Write data into table on SQLite database
}
