% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{round_df}
\alias{round_df}
\title{Round Numbers (2 digits behind comma)}
\usage{
round_df(x, digits = 2)
}
\arguments{
\item{x}{input}
}
\description{
Round Numbers (2 digits behind comma)
}
\examples{
round_df(100.1231212)
## [1] 100.12
}
