% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlite_dc}
\alias{sqlite_dc}
\title{Disconnect from SQLite database}
\usage{
sqlite_dc(path = getOption("mrsq_db_dir"),
  db_name = getOption("mrsq_db_name"))
}
\arguments{
\item{path}{folder path of the database}

\item{db_name}{name of the database}
}
\description{
Disconnect from SQLite database
}
