% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlite_drop}
\alias{sqlite_drop}
\title{Drop table on SQLite database}
\usage{
sqlite_drop(tbl_name, force = FALSE, path = getOption("mrsq_db_dir"),
  db_name = getOption("mrsq_db_name"))
}
\arguments{
\item{tbl_name}{name of the table}

\item{path}{folder path of the database}

\item{db_name}{name of the database}
}
\description{
Drop table on SQLite database
}
