% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlite_exec}
\alias{sqlite_exec}
\title{Execute query on SQLite database}
\usage{
sqlite_exec(query, path = getOption("mrsq_db_dir"),
  db_name = getOption("mrsq_db_name"))
}
\arguments{
\item{query}{query statement}

\item{path}{folder path of the database}

\item{db_name}{name of the database}
}
\description{
Execute query on SQLite database
}
