% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlite_glance}
\alias{sqlite_glance}
\title{Get first 10 rows of table on SQLite database}
\usage{
sqlite_glance(tbl_name, n = 10, path = getOption("mrsq_db_dir"),
  db_name = getOption("mrsq_db_name"))
}
\arguments{
\item{tbl_name}{name of the table}

\item{n}{number of limits}

\item{path}{folder path of the database}

\item{db_name}{name of the database}
}
\description{
Get first 10 rows of table on SQLite database
}
