% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlite_init}
\alias{sqlite_init}
\title{Initialize SQLite database}
\usage{
sqlite_init(path = getOption("mrsq_db_dir"),
  db_name = getOption("mrsq_db_name"))
}
\arguments{
\item{path}{folder path of the database}

\item{db_name}{name of the database}
}
\description{
Initialize SQLite database
}
