% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlite_query}
\alias{sqlite_query}
\title{Read table from SQLite database}
\usage{
sqlite_query(query, path = getOption("mrsq_db_dir"),
  db_name = getOption("mrsq_db_name"))
}
\arguments{
\item{query}{query statement}

\item{path}{folder path of the database}

\item{db_name}{name of the database}
}
\description{
Read table from SQLite database
}
