% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlite_start}
\alias{sqlite_start}
\title{Set SQLite database location and name}
\usage{
sqlite_start(path = "~/OneDrive/Magnum Opus/project/data/db",
  db_name = "db_qs.sqlite")
}
\arguments{
\item{path}{folder path of the database}

\item{db_name}{name of the database}
}
\description{
Set SQLite database location and name
}
