% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlite_table}
\alias{sqlite_table}
\title{Show tables on SQLite database}
\usage{
sqlite_table(pattern = NULL, path = getOption("mrsq_db_dir"),
  db_name = getOption("mrsq_db_name"))
}
\arguments{
\item{pattern}{regex pattern}

\item{path}{folder path of the database}

\item{db_name}{name of the database}
}
\description{
Show tables on SQLite database
}
