% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sqlite_write}
\alias{sqlite_write}
\title{Write data into table on SQLite database}
\usage{
sqlite_write(df, tbl_name, append = TRUE, overwrite = FALSE, date = TRUE,
  path = getOption("mrsq_db_dir"), db_name = getOption("mrsq_db_name"))
}
\arguments{
\item{df}{data.frame to insert}

\item{tbl_name}{name of the table}

\item{append}{TRUE}

\item{overwrite}{FALSE}

\item{date}{TRUE date convert to character}

\item{path}{folder path of the database}

\item{db_name}{name of the database}
}
\description{
Write data into table on SQLite database
}
