% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_din.R
\name{theme_din}
\alias{theme_din}
\title{ggplot2 theme using DIN font}
\usage{
theme_din(version = 1, panel_grid = TRUE, legend = NULL,
  plot_margin = margin(5, 5, 5, 5), font = "DIN", ...)
}
\arguments{
\item{version}{default is version 1 which is based on theme_ipsum,
version 2 is based on theme_minimal}

\item{panel_grid}{\code{TRUE}, \code{FALSE}}

\item{legend}{default is \code{NULL} (to fill, remember quadran position: 1, 2, 3, 4)}

\item{plot_margin}{margin(5, 5, 5, 5)}

\item{font}{default is DIN font}

\item{...}{other parameters}
}
\description{
ggplot2 theme using DIN font
}
